Moving admin functions to a new file. They are still part of the same pointer receiver type.
